Change compset in test
In an upcoming CTSM tag, the compset I2000Clm50BgcCropRtmGs will be
renamed to I2000Clm50BgcCropRtm (while keeping the same meaning as
before). This tag changes a test to use the new alias, to keep RTM
compatible with this upcoming CTSM tag.

Note that this change will require the upcoming CTSM tag (probably
ctsm5.1.dev011) in order to run this test from the rtm test list.